requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.0.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
pandas==2.3.1
numpy==2.3.1
openpyxl==3.1.5
python-dateutil==2.9.0
pytz==2025.2
stripe==14.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
